refactor file diff processor
https://coveord.atlassian.net/browse/CDX-347
